requirements.txt
fastapi
uvicorn
supabase
openai
python-dotenv
pydantic